move assertFreeBalance to utils
